requirements.txt
# ============================================
# Unified Requirements for CodeSlayer RFP Automation
# Consolidates: Sales-agent-main, Technichal_Agent-main, pricing_agent
# ============================================

# --- API & Server ---
fastapi==0.104.1
uvicorn==0.38.0
flask==3.1.2
flask-cors==6.0.1

# --- Core Python Dependencies ---
requests==2.32.5
beautifulsoup4==4.14.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
lxml==6.0.2
lxml_html_clean==0.4.3
tqdm==4.67.1
rich==14.2.0

# --- HTML Rendering / Scraping ---
requests-html==0.10.0
html5lib==1.1
bs4==0.0.2
pyquery==2.0.1
cssselect==1.3.0
fake-useragent==2.2.0

# --- AI & LLM Integrations ---
google-generativeai>=0.8.0
openai>=2.0.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.2.0
langgraph-checkpoint>=2.0.0
langsmith>=0.2.0

# --- Embeddings & Vector Search ---
sentence-transformers==5.1.2
chromadb==1.3.0
scikit-learn==1.7.2

# --- ML/AI Libraries ---
torch==2.9.0
transformers==4.57.1
tokenizers==0.22.1
numpy==2.3.4
scipy==1.16.3

# --- PDF Processing ---
pdfplumber==0.11.7
pdfminer.six==20250506
pypdf==6.1.3
pypdfium2==5.0.0

# --- File Handling & Data ---
pandas==2.3.3
openpyxl
jsonschema==4.25.1
PyYAML==6.0.3

# --- Data Validation ---
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4

# --- Environment & Utilities ---
typing-extensions==4.15.0
python-dotenv==1.2.1
jsonpatch==1.33
jsonpointer==3.0.0

